Added authoring tool link to the teacher homepage tabs and removed the '+ Create' button from the unit library tab
