+++ src/main/webapp/site/src/app/modules/library/teacher-project-library/teacher-project-library.component.html
@@ -1,9 +1,3 @@
-<div fxLayout="row" fxLayoutAlign="end center">
-  <a mat-raised-button color="accent" href="{{authoringToolLink}}">
-    <mat-icon>add</mat-icon>
-    <ng-container i18n>Create</ng-container>
-  </a>
-</div>
 <div class="library library-teacher notice-bg-bg" fxLayout="column" fxLayout.gt-sm="row">
   <div class="content-block dark-theme library__header library-teacher__header primary-bg"
        fxFlex.gt-sm="33">
