+++ src/main/webapp/site/src/app/modules/library/teacher-project-library/teacher-project-library.component.spec.ts
@@ -5,29 +5,9 @@ import { LibraryService } from "../../../services/library.service";
 import { fakeAsyncResponse } from "../../../student/student-run-list/student-run-list.component.spec";
 
 import { NO_ERRORS_SCHEMA } from "@angular/core";
-import { ConfigService } from "../../../services/config.service";
-import { Config } from "../../../domain/config";
-import { Observable } from 'rxjs';
 
 export class MockLibraryService {
   tabIndexSource$ = fakeAsyncResponse(1);
-}
-
-export class MockConfigService {
-  getConfig(): Observable<Config> {
-    const config: Config = {
-      contextPath: "/wise",
-      logOutURL: "/logout",
-      currentTime: "2018-10-17 00:00:00.0"
-    };
-    return Observable.create(observer => {
-      observer.next(config);
-      observer.complete();
-    });
-  }
-  getContextPath(): string {
-    return "";
-  }
 }
 
 describe('TeacherProjectLibraryComponent', () => {
@@ -39,8 +19,7 @@ describe('TeacherProjectLibraryComponent', () => {
       imports: [ MatMenuModule ],
       declarations: [ TeacherProjectLibraryComponent ],
       providers: [
-        { provide: LibraryService, useClass: MockLibraryService },
-        { provide: ConfigService, useClass: MockConfigService }
+        { provide: LibraryService, useClass: MockLibraryService }
       ],
       schemas: [ NO_ERRORS_SCHEMA ]
     })
