+++ src/main/webapp/site/src/app/teacher/teacher-home/teacher-home.component.html
@@ -7,12 +7,20 @@
   </h1>
 </header>
 <section>
-  <mat-tab-group [(selectedIndex)]="selectedTabIndex">
+  <mat-tab-group [(selectedIndex)]="selectedTabIndex" #tabs>
     <mat-tab label="Class Schedule">
       <app-teacher-run-list></app-teacher-run-list>
     </mat-tab>
     <mat-tab label="Browse WISE Units">
         <app-teacher-project-library></app-teacher-project-library>
+    </mat-tab>
+    <mat-tab>
+      <ng-template mat-tab-label>
+        <span fxLayout="row" fxLayoutAlign="center center">
+          <ng-container i18n>Authoring Tool</ng-container>
+          &nbsp;<mat-icon>open_in_new</mat-icon>
+        </span>
+      </ng-template>
     </mat-tab>
   </mat-tab-group>
 </section>
