+++ src/main/webapp/site/src/app/teacher/teacher-home/teacher-home.component.ts
@@ -1,7 +1,9 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, ViewChild } from '@angular/core';
 import { UserService } from "../../services/user.service";
 import { User } from "../../domain/user";
 import { TeacherService } from "../teacher.service";
+import { ConfigService } from "../../services/config.service";
+import { MatTab, MatTabGroup, MatTabHeader } from '@angular/material';
 
 @Component({
   selector: 'app-teacher-home',
@@ -9,12 +11,15 @@ import { TeacherService } from "../teacher.service";
   styleUrls: ['./teacher-home.component.scss']
 })
 export class TeacherHomeComponent implements OnInit {
+  @ViewChild('tabs') tabs: MatTabGroup;
 
   user: User = new User();
   selectedTabIndex: number = 0;
+  authoringToolLink: string = '';
 
   constructor(private userService: UserService,
-              private teacherService: TeacherService) {
+              private teacherService: TeacherService,
+              private configService: ConfigService) {
     teacherService.tabIndexSource$.subscribe((tabIndex) => {
       this.selectedTabIndex = tabIndex;
     });
@@ -22,6 +27,14 @@ export class TeacherHomeComponent implements OnInit {
 
   ngOnInit() {
     this.getUser();
+    this.configService.getConfig().subscribe((config) => {
+      if (config != null) {
+        this.authoringToolLink = `${this.configService.getContextPath()}/author`;
+      }
+    });
+    // Workaround for intercepting mat-tab change events
+    // https://stackoverflow.com/questions/51354135/how-to-conditionally-prevent-user-from-navigating-to-other-tab-in-mat-tab-group/51354403#51354403
+    this.tabs._handleClick = this.interceptTabChange.bind(this);
   }
 
   getUser() {
@@ -31,4 +44,11 @@ export class TeacherHomeComponent implements OnInit {
       });
   }
 
+  interceptTabChange(tab: MatTab, tabHeader: MatTabHeader, idx: number) {
+    if (idx === 2) {
+      window.location.href = this.authoringToolLink;
+      return false;
+    }
+    MatTabGroup.prototype._handleClick.apply(this.tabs, arguments);
+  }
 }
